Ease mmap requirement to increase windows compatibility
This makes it more compatible with running on Windows.
Note, though, that the main target for the decryptor is encrypted ESXi
images. Decrypting individual files on Windows is difficult, because we
need to know the plaintext of 64 encrypted bytes. This is feasible for
the ESXi files. For individual files, a lot of expertise is required to
make it work.